[AMDGPU] Support LLVM 24 and MI350
Adapt MLIR interfaces and MFMA declarations across LLVM versions, match the imported LLVM library linkage, and use the matching compiler for device bitcode. Detect gfx950 and use its OCP FP8 encodings in reference tests.

Validated the assembled stack with an LLVM24 rebuild, all 5 CTest targets, and 306 pytest tests plus 99 subtests covering JIT, memory primitives, FP4 GEMM and local/distributed MoE.

+++ lib/IR/mlir_generator_test.cc
@@ -1567,7 +1567,7 @@ def sched_group_barrier_test():
     bool found_sched_group_barrier = false;
     mlir->walk([&](mlir::ROCDL::SchedGroupBarrier op) {
         found_sched_group_barrier = true;
-        EXPECT_EQ(op.getMask(), 2u);
+        EXPECT_EQ(static_cast<uint32_t>(op.getMask()), 2u);
         EXPECT_EQ(op.getSize(), 4u);
         EXPECT_EQ(op.getGroupId(), 1u);
     });
@@ -1609,7 +1609,7 @@ def sched_barrier_test():
     bool found_sched_barrier = false;
     mlir->walk([&](mlir::ROCDL::SchedBarrier op) {
         found_sched_barrier = true;
-        EXPECT_EQ(op.getMask(), 0u);
+        EXPECT_EQ(static_cast<uint32_t>(op.getMask()), 0u);
     });
     EXPECT_TRUE(found_sched_barrier);
